Second attempt to prevent double-input on last note
The previous solution was tracking the `lastProcessedNoteRef`. The code
was updating this value to the next `targetNote` in the scale. This was
an issue because `targetNote` is fixed to a single octave, whereas the
note that is played through MIDI could be of any octave.

The new fix checks if the last processed note is in the `chordStack`. If
it is, we're probably detecting a duplicate input. For chords, we're
setting the last processed note to the first note detected in the chord
stack that matches the target chord stack.

I'm not completely satisfied with this solution, but it seems to work.